Move services and metrics_service command to tasks app
Refactor for better separation of concerns - CLI/management
functionality now lives with the tasks app it depends on.

Changes:
- Move apps/core/services/ to apps/tasks/services/
- Move metrics_service management command to apps/tasks/
- Update all imports in moved files and tests
- Move service tests to tests/unit/tasks/
