+++ apps/tasks/services/output_formatter.py
